CHE-370: add loader widget for the process of resolving dependencies

+++ plugins/plugin-maven/che-plugin-maven-ext-server/src/main/java/org/eclipse/che/ide/extension/maven/server/core/MavenTaskExecutor.java
@@ -76,7 +76,7 @@ public class MavenTaskExecutor {
                 restTasks = queue.size();
             }
 
-            notifier.setPercent(taskDone / (restTasks + taskDone));
+            notifier.setPercent((double)taskDone / (double)(restTasks + taskDone));
             try {
                 task.perform();
             } catch (Throwable throwable) {
